Login with web-wallet on every page
